Port Target Safety to the GraphQL API
Re-enabling the Safety section on the target page.

In the process, port the logic to format links based on Pubmed IDs from
the Angular webapp and make some of the list item keys passed to React
more robust (with the assumption that the same effect or publication
will not appear in the same row twice).